Revert "Automating Memory Leak Detection + Preventing a (Way Too Common) Async Bug"
