--- EssentialFeed/EssentialFeed/FeedAPI/HTTPClient.swift
@@ -1,17 +1,0 @@
-//
-//  HTTPClient.swift
-//  EssentialFeed
-//
-//  Created by Intan Nurjanah on 10/10/22.
-//
-
-import Foundation
-
-public enum HTTPClientResult {
-    case success(Data, HTTPURLResponse)
-    case failure(Error)
-}
-
-public protocol HTTPClient {
-    func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void)
-}
